Add button for toggling shutter movement
Home-Assistant doesn't remember the last movement direction and
just toggles between "not closed"/"closed". The duofern-internal
toggle behaves way better: when moving the shutter stops, when
stopped it moves in the opposite direction of the last movement.
This new entity calls this duofern-enabled toggle mechanism.
Since it doesn't have state, a button is the appropriate entity
type, even though there is no corresponding button on the actual
device.