Add switch for manual mode of duofern shutters
Duofern shutters can either move automatically or manually. In manual
mode they are never moved automatically (in non-manual mode, both is
fine). They have a button to toggle the manual mode, which is
exposed via the API. Since the mode can either be "on" or "off", the
correct entity in this case is a switch. The API exposes other
modes, which can be switched on or off, which could be easily extended
using the provided class. Some care would have to be taken around
adding or re-adding the entities, thouhg.