Rename id to duofernid for clarity
Around the entities there are multiple kinds of ids: duofern id,
unique (home assistant) id and the id that is used internally by
the integration (which mostly is the same as the duofern id, but
may have a suffix). By renaming the property it becomes clearer
which id is meant.