Expose switch for sun automatic in rollotrons
The sun automatic can close a cover based on light exposure on a
sensor that's stuck to the window. This can prevent overheating of
a room. However, in spring and fall (or generally when it's not too
hot outside) that actually might be desired. By exposing the switch
that toggles this behaviour (de-)activating it can be automated
in home assistant, e.g. by weather forecast.